backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
cryptography==42.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
bleach==6.1.0

# HTTP клієнт для API
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Monitoring
sentry-sdk[fastapi]==1.39.2
colorlog==6.8.2

# Testing (опціонально)
pytest==8.1.1
pytest-anyio
freezegun==1.5.0


# Для роботи з файлами
aiofiles==23.2.1
python-magic==0.4.27

# Email
resend==0.7.0
fastapi-mail==1.4.1

gunicorn